Configure netrw
Instead of installing a fuzzy finder or a project drawer, we'll use
ctags to switch between files, `:b` to switch between buffers, and Vim's
built-in netrw to browse directory structures.

To use ctags, install [exuberant ctags](http://ctags.sourceforge.net)
(`brew install ctags`) on macOS. Then create a tags file in your project
directory by running `ctags -R .` in the terminal, or `:! ctags -R .` in
Vim.

A useful tip is to ignore `node_modules` using the `--exclude` flag when
running the `ctags` utility (`ctags -R . --exclude=node_modules`), or by
creating a [`~/.ctags` file with global
configuration](https://gist.github.com/jeffkreeftmeijer/acf0d62232084942f2e86c9f890bf4fd).

With the tags generated, we can now place our cursor over a word, and
press `<C-n>` to jump to where the thing that word describes is defined.
To jump back up, we'll press `<C-t>`.

For switching between open files, we'll use `:b`. It takes a filename,
but a quick way to find the file you're looking for is `:b<space><tab>`,
which will show the list of all open buffers in a whildmenu.

Finally, netrw is Vim's built-in file browser. We'll update some of its
configuration to make it a bit nicer to work with.

    " Use netrw's tree view
    let g:netrw_liststyle = 3
    " Hide netrw's help banner (toggle with I)
    let g:netrw_banner = 0
    " Open files in the previous split
    let g:netrw_browse_split = 4
    " Make the netrw split use 20% of the window width
    let g:netrw_winsize = 20