Automatically run mix format and rustfmt on save
One of coc.nvim's features is to automatically format source files
through its plugins. To turn it on for both Elixir and Rust, open coc's
configuration file by calling `:CocConfig` in Vim, and add both to
`coc.preferences.formatOnSaveFiletypes`:

    {
      "coc.preferences.formatOnSaveFiletypes": ["elixir", "rust"]
    }

By default, the configuration file is stored in
`~/.config/nvim/coc-settings.json` when using NeoVim. To override this,
set `g:coc_config_home` in `~/.vimrc`:

    let g:coc_config_home = $HOME . "/.vim/"

After that, the configuration file that's loaded and opened with
`:CocConfig` is stored in `~/.vim/coc-settings.json`.